LPS-109769 Rename, following same pattern as autotagging
