+++ modules/apps/depot/depot-api/bnd.bnd
@@ -8,4 +8,4 @@ Export-Package:\
 	com.liferay.depot.model,\
 	com.liferay.depot.service,\
 	com.liferay.depot.service.persistence,\
-	com.liferay.depot.util
+	com.liferay.depot.configuration
