+++ modules/apps/depot/depot-api/src/main/resources/com/liferay/depot/configuration/packageinfo
