+++ modules/apps/depot/depot-web/src/main/java/com/liferay/depot/web/internal/application/list/DepotSettingsPanelApp.java
@@ -17,7 +17,7 @@ package com.liferay.depot.web.internal.application.list;
 import com.liferay.application.list.BasePanelApp;
 import com.liferay.application.list.PanelApp;
 import com.liferay.application.list.constants.PanelCategoryKeys;
-import com.liferay.depot.util.DepotSupportChecker;
+import com.liferay.depot.configuration.DepotConfiguration;
 import com.liferay.depot.web.internal.constants.DepotPortletKeys;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.Group;
@@ -52,7 +52,7 @@ public class DepotSettingsPanelApp extends BasePanelApp {
 		throws PortalException {
 
 		if ((group.getType() != GroupConstants.TYPE_DEPOT) ||
-			!_depotSupportChecker.isEnabled()) {
+			!_depotConfiguration.isEnabled()) {
 
 			return false;
 		}
@@ -70,7 +70,7 @@ public class DepotSettingsPanelApp extends BasePanelApp {
 	}
 
 	@Reference
-	private DepotSupportChecker _depotSupportChecker;
+	private DepotConfiguration _depotConfiguration;
 
 	@Reference
 	private Portal _portal;
