+++ modules/apps/depot/depot-web/src/main/java/com/liferay/depot/web/internal/item/selector/provider/DepotGroupItemSelectorProvider.java
@@ -16,7 +16,7 @@ package com.liferay.depot.web.internal.item.selector.provider;
 
 import com.liferay.depot.model.DepotEntry;
 import com.liferay.depot.service.DepotEntryService;
-import com.liferay.depot.util.DepotSupportChecker;
+import com.liferay.depot.configuration.DepotConfiguration;
 import com.liferay.item.selector.provider.GroupItemSelectorProvider;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.Language;
@@ -111,7 +111,7 @@ public class DepotGroupItemSelectorProvider
 
 	@Override
 	public boolean isEnabled() {
-		return _depotSupportChecker.isEnabled();
+		return _depotConfiguration.isEnabled();
 	}
 
 	private long _getLiveGroupId(long groupId) throws PortalException {
@@ -131,7 +131,7 @@ public class DepotGroupItemSelectorProvider
 	private DepotEntryService _depotEntryService;
 
 	@Reference
-	private DepotSupportChecker _depotSupportChecker;
+	private DepotConfiguration _depotConfiguration;
 
 	@Reference
 	private GroupService _groupService;
